requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
dj-database-url==0.5.0
Django==3.0
django-debug-toolbar==3.2.2
django-heroku==0.0.0
frozenlist==1.2.0
fsspec==2021.11.1
future==0.18.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.42.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.2
install==1.3.5
joblib==1.1.0
Markdown==3.3.6
multidict==5.2.0
numpy==1.21.4
oauthlib==3.1.1
packaging==21.3
Pillow==8.4.0
protobuf==3.19.1
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pyparsing==3.0.6
pytorch-lightning==1.5.5
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
sqlparse==0.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==3.0.0
torch==1.10.0
torchmetrics==0.6.1
torchvision==0.11.1
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
whitenoise==5.3.0
yarl==1.7.2
zipp==3.6.0
